+++ example/lib/widgets/demo_page_scaffold.dart
@@ -19,6 +19,9 @@ class DemoPageScaffold extends StatelessWidget {
   Widget build(BuildContext context) {
     return LiquidGlassBackdropGroup(
       disableBlurWhileScrolling: false,
+      settings: LiquidGlassSettings(
+          androidBlurSigma: 4
+      ),
       child: SingleChildScrollView(
         padding: const EdgeInsets.fromLTRB(24, 32, 24, 120),
         child: Column(